+++ aa-display/src/main/java/io/github/nitsuya/aa/display/xposed/CoreManagerService.kt
@@ -271,10 +271,6 @@ class CoreManagerService private constructor() : ICoreManager.Stub() {
             }?.trim()?.takeIf { it.isNotEmpty() }
         }
 
-        fun buriedPackagesOnAaDisplays(): Set<String> {
-            return mSplitController?.buriedPackagesOnAaDisplays().orEmpty()
-        }
-
         fun avStackLayout(): AvMediaArbiter.StackLayout? {
             return mSplitController?.avStackLayout()
         }